src/test/java/com/codingbat/array3/MaxMirrorTest.java
/*
  CodingBat Array 3 - https://codingbat.com/java/Array-3

  My repository with solutions - https://github.com/egalli64/jex/
 */
package com.codingbat.array3;

import static org.assertj.core.api.Assertions.assertThat;

import java.util.stream.Stream;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class MaxMirrorTest {
    static Stream<Arguments> provider() {
        return Stream.of( //
                Arguments.of(new int[] { 1, 2, 3, 8, 9, 3, 2, 1 }, 3), //
                Arguments.of(new int[] { 1, 2, 1, 4 }, 3), //
                Arguments.of(new int[] { 7, 1, 2, 9, 7, 2, 1 }, 2) //
        );
    }

    @ParameterizedTest
    @MethodSource("provider")
    void solutionExamples(int[] input, int expected) {
        assertThat(MaxMirror.solution(input)).isEqualTo(expected);
    }
}
